news_scrape/requirements.txt
httpx==0.27.1
aiohttp==3.9.3
cloudscraper>=1.2.71
curl-cffi>=0.9
beautifulsoup4==4.12.3
lxml>=4.9.4
pendulum>=3.0.0
trafilatura==1.8.1
urllib3==1.26.18
# MongoDB
motor==3.6
pymongo>=4.9

spacy==3.7.2
thinc==8.2.1
numpy==1.26.4
pandas==2.1.4
# Environment and configuration
python-dotenv==1.0.0

# API Framework
fastapi==0.103.1
uvicorn==0.23.2

# Airflow
apache-airflow==2.10.2

# Utilities
python-json-logger==2.0.7
pydantic==2.10.6